requirements.txt
python-dotenv==0.17.1
python-lyric==1.1.6
influxdb-client==1.17.0